+++ src/lib-storage/index/mbox/mbox-storage.c
@@ -1,4 +1,4 @@
-/* Copyright (c) 2002-2016 Dovecot authors, see the included COPYING file */
+/* Copyright (c) 2002-2017 Dovecot authors, see the included COPYING file */
 
 #include "lib.h"
 #include "ioloop.h"
@@ -689,7 +689,7 @@ static bool
 mbox_is_internal_name(struct mailbox_list *list ATTR_UNUSED,
 		      const char *name)
 {
-	unsigned int len;
+	size_t len;
 
 	/* don't allow *.lock files/dirs */
 	len = strlen(name);
